% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rwPFS.R
\name{calc_rwPFS}
\alias{calc_rwPFS}
\title{Calculate the real-world Progression-free Survival (rwPFS) endpoint}
\usage{
calc_rwPFS(
  .df,
  .start_date = NULL,
  .visit_gap_start_date = NULL,
  .last_progression_abstraction_date = NULL,
  .progression_date = NULL,
  .last_activity_date = NULL,
  .death_date = NULL,
  .death_window_days = NULL,
  .max_follow_up_days = Inf,
  .label = ""
)
}
\arguments{
\item{.df}{a data frame}

\item{.start_date}{character. The name of a date column in .df that represents
the baseline date}

\item{.visit_gap_start_date}{character. The name of a date column in .df that represents
the date of the last visit before a large (usually >90d) visit gap occurring after start_date (if any).
Progression follow-up will be censored at this date (latest). Dates should be NA if there's no visit gap.}

\item{.last_progression_abstraction_date}{character. The name of a date column in .df that represents
the date up which real-world progression was abstracted in the database.
Progression follow-up will be censored at this date (latest).}

\item{.progression_date}{character. The name of a date column in .df that represents
the date of first progression after start date (if any). Dates should be NA if there's no progression.}

\item{.last_activity_date}{character. The name of a date column in .df that represents
the date of last structural activity in the database. Progression follow-up will be censored at this date (latest).}

\item{.death_date}{character. The name of a date column in .df that represents the date of death.
Dates should be NA if there's no recorded death.}

\item{.death_window_days}{integer. How many days after end of progression follow-up should death events be included?
This is necessary because patients often drop out of the database shortly before death, and (only) death events correlated with
the end of follow-up should be captured. Common values are 30-60 days, but this may
depend on various factors. See vignette for how to decide on the length of this time window.}

\item{.max_follow_up_days}{integer. Maximum number of days after which patients will be censored.}

\item{.label}{character. Label to keep track of multiple rwPFS endpoints in one dataset.}
}
\value{
Returns a data frame with the following new data columns added:

\strong{New Columns:}

\describe{

\item{rwpfs<label>_eof_date}{The end date of progression follow-up. This is the earliest non-missing date of \code{.visit_gap_start_date},
\code{.last_progression_abstraction_date}, and \code{.last_activity_date}}

\item{rwpfs<label>_event_type}{The type of rwPFS event: "Death", "Progression", "Censored", or "Missing" (if rwPFS could not be calculated)}

\item{rwpfs<label>_date}{The rwPFS event- or censoring date.}

\item{rwpfs<label>_days}{The time from \code{.start_date} to \code{rwpfs<label>_date} in days.}

\item{rwpfs<label>_event}{Whether a rwPFS event was recorded (1) or the observation was censored (0) }

\item{rwpfs<label>_months}{The time from \code{.start_date} to \code{rwpfs<label>_date} in months.}

}

If rwPFS could not be calculated, the \code{rwpfs<label>_event_type} will be set to "Missing", and all other columns will contain missing values.
This can be due to one of the following reasons: missing \code{.last_progression_abstraction_date} <= \code{.start_date}, missing \code{.last_activity_date} or <= \code{.start_date}, \code{.last_progression_abstraction_date} <= \code{.start_date},
or \code{.death_date} <= \code{.start_date} (this can happen when death dates are rounded for anonymization purposes).
}
\description{
Calculates the real-world PFS (rwPFS) endpoint given information on follow-up period, progression-, and death dates.
}
\examples{

\dontrun{


library(dplyr)
library(RwPFS) 

#Starting point: the simprog simulated dataset (included in RwPFS) 

#Add rwPFS:
df <- simprog \%>\% 
  calc_rwPFS(
    .start_date = "baseline_date",                    #baseline date for measuring time-to-event
    .visit_gap_start_date = "visit_gap_start_date",   #the start date of any gap in visits >90 days  
    .last_progression_abstraction_date = "last_progression_abstraction_date", 
    .progression_date = "progression_date_all_events",#the date of progression (none: NA)
    .last_activity_date = "last_activity_date",       #the date of last activity in the database
    .death_date = "death_date",                       #the date of death
    .death_window_days = 30,                          #include death events <30d after progression EOF
    .max_follow_up_days = Inf,                        #censor patients after a maximum time. 
    .label = "_all_events"                            #Label for this rwPFS endpoint (for comparisons)
  )

#View the result:
df \%>\% 
  select(contains("rwPFS")) \%>\% 
  glimpse()

}


}
