% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simprog}
\alias{simprog}
\title{Simulated progression data}
\format{
A data frame with simulated patient data for PFS analysis
}
\usage{
simprog
}
\description{
A demo dataset containing simulated data to compute progression-free
survival. For demonstration purposes only.
}
\examples{
data(simprog)
head(simprog)
}
\keyword{datasets}
