% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rwPFS.R
\name{compare_rwPFS}
\alias{compare_rwPFS}
\title{Compare different real-world PFS (rwPFS) definitions}
\usage{
compare_rwPFS(
  .df,
  .labels = NULL,
  .reference = NULL,
  .incremental_deaths_column = NULL
)
}
\arguments{
\item{.df}{a data frame}

\item{.labels}{character vector. A vector of strings representing the labels that identify the
rwPFS endpoints/definitions to be compared (in order of desired appearance in the output table).}

\item{.reference}{character. Label identifying the reference rwPFS definition that others are compared against.}

\item{.incremental_deaths_column}{logical. Should the output table contain an additional column comparing the incremental
number of death events captured from one definition to the next? Intended for comparing different \code{death_window_days}
parameter values in \code{fi_calc_rwPFS}.}
}
\value{
Returns a summary table
}
\description{
Compares a range of different rwPFS definitions with respect to event composition, KM median time to event,
Hazard ratio relative to the reference, and incremental no. of death events from one rwPFS definition to the next (optional).
}
\examples{

#'\dontrun{

#Note: the FlatironKitchen package is used in these
#examples for simplicity. This is not a requirement.

#Initialize FlatironKitchen object
fk <- fi_start(datamart = "AdvancedNSCLC",
               title = "rwPFS in aNSCLC")  \%>\%
  
  #Use start of line as start_date
  fi_add_lineoftherapy_flatiron(
    lines = c(1, 2),
    index_date = "advanceddiagnosisdate",
    left = 0,
    right = 90,
    calc_duration = FALSE
  ) \%>\%
  
  #Restrict to Carboplatin & Paclitaxel in 1st line
  fi_cohort_include(
    lot1linename == "Carboplatin,Paclitaxel",
    description = "Carboplatin & Paclitaxel in 1st line",
    keep_na = FALSE
  )  \%>\%
  
  #Add information on any long visit gaps
  fi_calc_visitgap(
    index_date = "lot1startdate",
    gapdays = 90,
    what = "After",
    force_database = FALSE
  )

#Pull the analysis dataset from 
#the FlatironKitchen object
df <- fk \%>\%
  fi_pull_data()

#Download the progression table belonging
#to the current datamart
progression <- fk \%>\%
  fi_read_table(table_name="ENHANCED_ADVNSCLCPROGRESSION") \%>\%
  collect()

#Add progression information to analysis dataset
df <- df \%>\%
  #select the start_date column..
  select(patientid, lot1startdate) \%>\%
  
  #..and add it to the raw progression table
  left_join(progression, by = "patientid") \%>\%
  
  #Aggregate progression information (one-row per patient)
  filter_progression(
    .start_date = "lot1startdate",
    .require_radiographic = FALSE,
    .exclude_pseudoprogression = TRUE,
    .discard_n_days = 0,
    .label = "_no_pseudo_no_mixed",
    .prog_filter_expression = "ismixedresponse == 'No'"
  ) \%>\%
  
  #Combine progression columns with analysis dataset
  right_join(df, by = "patientid")


#Add several different rwPFS endpoints
df <- df \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 0,
    .max_follow_up_days = 1200,
    .label = "_0d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 10,
    .max_follow_up_days = 1200,
    .label = "_10d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 20,
    .max_follow_up_days = 1200,
    .label = "_20d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 30,
    .max_follow_up_days = 1200,
    .label = "_30d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 40,
    .max_follow_up_days = 1200,
    .label = "_40d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 50,
    .max_follow_up_days = 1200,
    .label = "_50d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 60,
    .max_follow_up_days = 1200,
    .label = "_60d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 70,
    .max_follow_up_days = 1200,
    .label = "_70d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 80,
    .max_follow_up_days = 1200,
    .label = "_80d_window"
  ) \%>\%
  
  calc_rwPFS(
    .start_date = "lot1startdate",
    .visit_gap_start_date = "lastvisitbeforegap",
    .last_progression_abstraction_date = "lastclinicnotedate_no_pseudo_no_mixed",
    .progression_date = "progressiondate_no_pseudo_no_mixed",
    .last_activity_date = "lastactivitydate",
    .death_date = "dateofdeath",
    .death_window_days = 90,
    .max_follow_up_days = 1200,
    .label = "_90d_window"
  ) \%>\%
  
  #Remove patients with missing rwPFS
  filter(
    rwPFS_0d_window_event_type != "Missing"
  )

#Compare different rwPFS definitions
df \%>\%
  compare_rwPFS(
    .labels = paste0("_", 0:9*10, "d_window"),
    .reference = "_30d_window",
    .incremental_deaths_column = TRUE
  ) \%>\%
  knitr::kable()
}

}
